requirements.txt
alembic==1.4.3
asgiref==3.2.10
Babel==2.9.1
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.1
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-hCaptcha==0.5.4
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mailman==0.1.5
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google-api-core==1.31.0
google-api-python-client==2.13.0
google-auth==1.33.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
limits==1.5.1
mailjet-rest==1.3.4
Mako==1.1.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sentry-sdk==0.19.2
six==1.16.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
